Cleaned up arg_width and arg_column functions. Added XXn and XXm support to -offset (not documented in mdoc.samples, but used in mdoc.samples!). Lint warnings fixed.

+++ term.c
@@ -315,7 +315,7 @@ term_flushln(struct termp *p)
 				putchar(' ');
 		}
 		if (TERMP_HANG & p->flags)
-			sv = vis - maxvis;
+			sv = (int)(vis - maxvis);
 		p->col = 0;
 		return;
 	}
